+++ tests/integ_tests/workflow_tests/activities.rs
@@ -1,7 +1,6 @@
-use crate::integ_tests::{init_core_and_create_wf, schedule_activity_cmd, CoreWfStarter};
+use crate::integ_tests::{init_core_and_create_wf, schedule_activity_cmd};
 use assert_matches::assert_matches;
 use std::time::Duration;
-use temporal_sdk_core::test_workflow_driver::{CommandSender, TestWorkflowDriver};
 use temporal_sdk_core::{
     protos::coresdk::{
         activity_result::{self, activity_result as act_res, ActivityResult},
@@ -9,8 +8,7 @@ use temporal_sdk_core::{
         common::{Payload, UserCodeFailure},
         workflow_activation::{wf_activation_job, FireTimer, ResolveActivity, WfActivationJob},
         workflow_commands::{
-            ActivityCancellationType, CompleteWorkflowExecution, RequestCancelActivity,
-            ScheduleActivity, StartTimer,
+            ActivityCancellationType, CompleteWorkflowExecution, RequestCancelActivity, StartTimer,
         },
         workflow_completion::WfActivationCompletion,
         ActivityTaskCompletion,
@@ -18,67 +16,6 @@ use temporal_sdk_core::{
     IntoCompletion,
 };
 use tokio::time::sleep;
-
-#[tokio::test(flavor = "multi_thread")]
-async fn activity_workflow_new() {
-    let wf_name = "activity_wf_new";
-    let mut starter = CoreWfStarter::new(wf_name);
-    let mut worker = starter.worker().await;
-
-    let activity_id = "act-1";
-    let activity_timeout = Duration::from_secs(5);
-    let payload_dat = b"hello".to_vec();
-
-    let twd = TestWorkflowDriver::new(|mut command_sink: CommandSender| {
-        let task_queue = starter.get_task_queue().to_owned();
-        let payload_dat = payload_dat.clone();
-        async move {
-            let activity = ScheduleActivity {
-                activity_id: activity_id.to_string(),
-                activity_type: "test_activity".to_string(),
-                task_queue: task_queue.clone(),
-                schedule_to_start_timeout: Some(activity_timeout.into()),
-                start_to_close_timeout: Some(activity_timeout.into()),
-                schedule_to_close_timeout: Some(activity_timeout.into()),
-                heartbeat_timeout: Some(activity_timeout.into()),
-                cancellation_type: ActivityCancellationType::TryCancel as i32,
-                ..Default::default()
-            };
-            let res = command_sink
-                .activity(activity)
-                .await
-                .unwrap()
-                .unwrap_ok_payload();
-            assert_eq!(res.data, payload_dat);
-            command_sink.complete_workflow_execution();
-        }
-    });
-    worker.submit_wf(wf_name.to_owned(), twd).await.unwrap();
-
-    let core = starter.get_core().await;
-    tokio::join! {
-        async {
-            worker.run_until_done().await.unwrap();
-        },
-        async {
-            // Poll activity and verify that it's been scheduled with correct parameters
-            let task = core.poll_activity_task().await.unwrap();
-            assert_matches!(
-                task.variant,
-                Some(act_task::Variant::Start(start_activity)) => {
-                    assert_eq!(start_activity.activity_type, "test_activity".to_string())
-                }
-            );
-            // Complete activity successfully.
-            core.complete_activity_task(ActivityTaskCompletion {
-                task_token: task.task_token,
-                result: Some(ActivityResult::ok(payload_dat.into())),
-            })
-            .await
-            .unwrap();
-        }
-    };
-}
 
 #[tokio::test]
 async fn activity_workflow() {
