+++ tests/main.rs
@@ -1,6 +1,7 @@
 #[cfg(test)]
 mod integ_tests {
     mod heartbeat_tests;
+    mod load_tests;
     mod polling_tests;
     mod workflow_tests;
